requirements.txt
annotated-types==0.7.0
b2sdk==2.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
idna==3.10
logfury==1.0.1
mysql-connector-python==9.3.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
